# src/main/resources/api.yml
openapi: 3.0.3
info:
  title: Transaction Microservice API
  description: API documentation for Transaction microservice.
  version: 1.0.0
servers:
  - url: http://localhost:8085/api
    description: Local development server
paths:
  /transactions/deposit:
    post:
      summary: Deposit funds to a bank account
      description: Creates a deposit transaction and updates the balance of the specified account.
      tags:
        - Transactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DepositTransaction'
      responses:
        '201':
          description: Deposit transaction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DepositResponse'
        '400':
          description: Invalid request
        '404':
          description: Account not found
        '500':
          description: Internal server error
  /transactions/withdraw:
    post:
      summary: Withdraw funds from a bank account
      description: Creates a withdrawal transaction and updates the balance of the specified account.
      tags:
        - Transactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/WithdrawTransaction'
      responses:
        '201':
          description: Withdrawal transaction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/WithdrawResponse'
        '400':
          description: Invalid request or insufficient balance
        '404':
          description: Account not found
        '500':
          description: Internal server error
  /transactions/transfer:
    post:
      summary: Transfer funds between two bank accounts
      description: Creates a transfer transaction and updates the balances of the source and destination accounts.
      tags:
        - Transactions
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TransferTransaction'
      responses:
        '201':
          description: Transfer transaction created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TransferResponse'
        '400':
          description: Invalid request or insufficient balance
        '404':
          description: Account not found
        '500':
          description: Internal server error
  /transactions:
    get:
      summary: Retrieve global transaction history
      description: Fetches all transactions, sorted by date in descending order.
      tags:
        - Transactions
      responses:
        '200':
          description: List of all transactions
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'
        '500':
          description: Internal server error
  /transactions/account/{accountId}:
    get:
      summary: Retrieve transaction history for a specific account
      description: Fetches transactions where the account is either the source or destination.
      tags:
        - Transactions
      parameters:
        - name: accountId
          in: path
          required: true
          description: The ID of the account whose transactions are to be retrieved.
          schema:
            type: string
      responses:
        '200':
          description: List of transactions for the specified account
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/TransactionResponse'
        '404':
          description: Account not found
        '500':
          description: Internal server error
components:
  schemas:
    DepositTransaction:
      type: object
      properties:
        accountId:
          type: string
          description: The ID of the account to deposit funds into.
          example: "1"
        amount:
          type: number
          format: double
          description: The amount to deposit.
          example: 500.0
        type:
          type: string
          enum:
            - DEPOSIT
          description: Type of transaction.
          example: "DEPOSIT"

    WithdrawTransaction:
      type: object
      properties:
        accountId:
          type: string
          description: The ID of the account to withdraw funds from.
          example: "1"
        amount:
          type: number
          format: double
          description: The amount to withdraw.
          example: 200.0
        type:
          type: string
          enum:
            - WITHDRAWAL
          description: Type of transaction.
          example: "WITHDRAWAL"

    TransferTransaction:
      type: object
      properties:
        sourceAccountId:
          type: string
          description: The ID of the account to transfer funds from.
          example: "1"
        destinationAccountId:
          type: string
          description: The ID of the account to transfer funds to.
          example: "2"
        amount:
          type: number
          format: double
          description: The amount to transfer.
          example: 300.0
        type:
          type: string
          enum:
            - TRANSFER
          description: Type of transaction.
          example: "TRANSFER"

    DepositResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the deposit transaction.
          example: "123"
        type:
          type: string
          enum:
            - DEPOSIT
          description: Type of transaction.
          example: "DEPOSIT"
        amount:
          type: number
          format: double
          description: The amount deposited.
          example: 500.0
        date:
          type: string
          format: date-time
          description: Date and time of the deposit transaction.
          example: "2024-11-21T08:00:00Z"
        destinationAccountId:
          type: string
          description: The account where the funds were deposited.
          example: "1"

    WithdrawResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the withdrawal transaction.
          example: "456"
        type:
          type: string
          enum:
            - WITHDRAWAL
          description: Type of transaction.
          example: "WITHDRAWAL"
        amount:
          type: number
          format: double
          description: The amount withdrawn.
          example: 200.0
        date:
          type: string
          format: date-time
          description: Date and time of the withdrawal transaction.
          example: "2024-11-21T09:00:00Z"
        sourceAccountId:
          type: string
          description: The account from which the funds were withdrawn.
          example: "1"

    TransferResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the transfer transaction.
          example: "789"
        type:
          type: string
          enum:
            - TRANSFER
          description: Type of transaction.
          example: "TRANSFER"
        amount:
          type: number
          format: double
          description: The amount transferred.
          example: 300.0
        date:
          type: string
          format: date-time
          description: Date and time of the transfer transaction.
          example: "2024-11-21T10:00:00Z"
        sourceAccountId:
          type: string
          description: The account from which the funds were transferred.
          example: "1"
        destinationAccountId:
          type: string
          description: The account to which the funds were transferred.
          example: "2"

    TransactionResponse:
      type: object
      properties:
        id:
          type: string
          description: Unique identifier for the transaction.
          example: "1"
        type:
          type: string
          enum:
            - DEPOSIT
            - WITHDRAWAL
            - TRANSFER
          description: Type of transaction.
          example: "DEPOSIT"
        amount:
          type: number
          format: double
          description: Amount involved in the transaction.
          example: 500.0
        date:
          type: string
          format: date-time
          description: Date and time of the transaction.
          example: "2024-11-20T08:00:00Z"
        sourceAccountId:
          type: string
          description: Originating account ID (only for withdrawal or transfer).
          example: "1"
        destinationAccountId:
          type: string
          description: Destination account ID (only for transfer).
          example: "2"